requirements.txt
# requirements.txt

# Simulation framework
cocotb>=1.9

# Test runner
pytest>=7.4

# Project manifest parsing
# Required by framework/project.py for auto-discovery of projects/*/project.yaml
pyyaml>=6.0
